requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
coinbase-commerce==1.0.1
Django==4.1.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-writable-nested==0.7.0
ecdsa==0.18.0
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.1.1
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
